Move DbError to scylla-cql-core
