+++ scylla-cql/src/frame/response/error.rs
@@ -1,12 +1,13 @@
 //! CQL protocol-level representation of an `ERROR` response.
 
-use crate::Consistency;
 use crate::frame::frame_errors::{CqlErrorParseError, LowLevelDeserializationError};
 use crate::frame::protocol_features::ProtocolFeatures;
 use crate::frame::types;
 use byteorder::ReadBytesExt;
 use bytes::Bytes;
-use thiserror::Error;
+
+// Re-exports from scylla-cql-core for backward compatibility.
+pub use scylla_cql_core::frame::response::error::{DbError, OperationType, WriteType};
 
 /// Represents a CQL protocol-level error that is sent by the database in response to a query
 /// that failed to execute successfully.
@@ -155,296 +156,6 @@ impl Error {
         Ok(Error { error, reason })
     }
 }
-
-/// An error sent from the database in response to a query
-/// as described in the [specification](https://github.com/apache/cassandra/blob/5ed5e84613ef0e9664a774493db7d2604e3596e0/doc/native_protocol_v4.spec#L1029)\
-#[derive(Error, Debug, Clone, PartialEq, Eq)]
-#[non_exhaustive]
-pub enum DbError {
-    /// The submitted query has a syntax error
-    #[error("The submitted query has a syntax error")]
-    SyntaxError,
-
-    /// The query is syntactically correct but invalid
-    #[error("The query is syntactically correct but invalid")]
-    Invalid,
-
-    /// Attempted to create a keyspace or a table that was already existing
-    #[error(
-        "Attempted to create a keyspace or a table that was already existing \
-        (keyspace: {keyspace}, table: {table})"
-    )]
-    AlreadyExists {
-        /// Created keyspace name or name of the keyspace in which table was created
-        keyspace: String,
-        /// Name of the table created, in case of keyspace creation it's an empty string
-        table: String,
-    },
-
-    /// User defined function failed during execution
-    #[error(
-        "User defined function failed during execution \
-        (keyspace: {keyspace}, function: {function}, arg_types: {arg_types:?})"
-    )]
-    FunctionFailure {
-        /// Keyspace of the failed function
-        keyspace: String,
-        /// Name of the failed function
-        function: String,
-        /// Types of arguments passed to the function
-        arg_types: Vec<String>,
-    },
-
-    /// Authentication failed - bad credentials
-    #[error("Authentication failed - bad credentials")]
-    AuthenticationError,
-
-    /// The logged user doesn't have the right to perform the query
-    #[error("The logged user doesn't have the right to perform the query")]
-    Unauthorized,
-
-    /// The query is invalid because of some configuration issue
-    #[error("The query is invalid because of some configuration issue")]
-    ConfigError,
-
-    /// Not enough nodes are alive to satisfy required consistency level
-    #[error(
-        "Not enough nodes are alive to satisfy required consistency level \
-        (consistency: {consistency}, required: {required}, alive: {alive})"
-    )]
-    Unavailable {
-        /// Consistency level of the query
-        consistency: Consistency,
-        /// Number of nodes required to be alive to satisfy required consistency level
-        required: i32,
-        /// Found number of active nodes
-        alive: i32,
-    },
-
-    /// The request cannot be processed because the coordinator node is overloaded
-    #[error("The request cannot be processed because the coordinator node is overloaded")]
-    Overloaded,
-
-    /// The coordinator node is still bootstrapping
-    #[error("The coordinator node is still bootstrapping")]
-    IsBootstrapping,
-
-    /// Error during truncate operation
-    #[error("Error during truncate operation")]
-    TruncateError,
-
-    /// Not enough nodes responded to the read request in time to satisfy required consistency level
-    #[error(
-        "Not enough nodes responded to the read request in time to satisfy required consistency level \
-            (consistency: {consistency}, received: {received}, required: {required}, data_present: {data_present})"
-    )]
-    ReadTimeout {
-        /// Consistency level of the query
-        consistency: Consistency,
-        /// Number of nodes that responded to the read request
-        received: i32,
-        /// Number of nodes required to respond to satisfy required consistency level
-        required: i32,
-        /// Replica that was asked for data has responded
-        data_present: bool,
-    },
-
-    /// Not enough nodes responded to the write request in time to satisfy required consistency level
-    #[error(
-        "Not enough nodes responded to the write request in time to satisfy required consistency level \
-            (consistency: {consistency}, received: {received}, required: {required}, write_type: {write_type})"
-    )]
-    WriteTimeout {
-        /// Consistency level of the query
-        consistency: Consistency,
-        /// Number of nodes that responded to the write request
-        received: i32,
-        /// Number of nodes required to respond to satisfy required consistency level
-        required: i32,
-        /// Type of write operation requested
-        write_type: WriteType,
-    },
-
-    /// A non-timeout error during a read request
-    #[error(
-        "A non-timeout error during a read request \
-        (consistency: {consistency}, received: {received}, required: {required}, \
-        numfailures: {numfailures}, data_present: {data_present})"
-    )]
-    ReadFailure {
-        /// Consistency level of the query
-        consistency: Consistency,
-        /// Number of nodes that responded to the read request
-        received: i32,
-        /// Number of nodes required to respond to satisfy required consistency level
-        required: i32,
-        /// Number of nodes that experience a failure while executing the request
-        numfailures: i32,
-        /// Replica that was asked for data has responded
-        data_present: bool,
-    },
-
-    /// A non-timeout error during a write request
-    #[error(
-        "A non-timeout error during a write request \
-        (consistency: {consistency}, received: {received}, required: {required}, \
-        numfailures: {numfailures}, write_type: {write_type}"
-    )]
-    WriteFailure {
-        /// Consistency level of the query
-        consistency: Consistency,
-        /// Number of nodes that responded to the read request
-        received: i32,
-        /// Number of nodes required to respond to satisfy required consistency level
-        required: i32,
-        /// Number of nodes that experience a failure while executing the request
-        numfailures: i32,
-        /// Type of write operation requested
-        write_type: WriteType,
-    },
-
-    /// Tried to execute a prepared statement that is not prepared. Driver should prepare it again
-    #[error(
-        "Tried to execute a prepared statement that is not prepared. Driver should prepare it again"
-    )]
-    Unprepared {
-        /// Statement id of the requested prepared query
-        statement_id: Bytes,
-    },
-
-    /// Internal server error. This indicates a server-side bug
-    #[error("Internal server error. This indicates a server-side bug")]
-    ServerError,
-
-    /// Invalid protocol message received from the driver
-    #[error("Invalid protocol message received from the driver")]
-    ProtocolError,
-
-    /// Rate limit was exceeded for a partition affected by the request.
-    /// (Scylla-specific)
-    /// TODO: Should this have a "Scylla" prefix?
-    #[error("Rate limit was exceeded for a partition affected by the request")]
-    RateLimitReached {
-        /// Type of the operation rejected by rate limiting.
-        op_type: OperationType,
-        /// Whether the operation was rate limited on the coordinator or not.
-        /// Writes rejected on the coordinator are guaranteed not to be applied
-        /// on any replica.
-        rejected_by_coordinator: bool,
-    },
-
-    /// Other error code not specified in the specification
-    #[error("Other error not specified in the specification. Error code: {0}")]
-    Other(i32),
-}
-
-impl DbError {
-    /// Returns the error code for this error, as defined in the CQL protocol specification.
-    pub fn code(&self, protocol_features: &ProtocolFeatures) -> i32 {
-        match self {
-            DbError::ServerError => 0x0000,
-            DbError::ProtocolError => 0x000A,
-            DbError::AuthenticationError => 0x0100,
-            DbError::Unavailable {
-                consistency: _,
-                required: _,
-                alive: _,
-            } => 0x1000,
-            DbError::Overloaded => 0x1001,
-            DbError::IsBootstrapping => 0x1002,
-            DbError::TruncateError => 0x1003,
-            DbError::WriteTimeout {
-                consistency: _,
-                received: _,
-                required: _,
-                write_type: _,
-            } => 0x1100,
-            DbError::ReadTimeout {
-                consistency: _,
-                received: _,
-                required: _,
-                data_present: _,
-            } => 0x1200,
-            DbError::ReadFailure {
-                consistency: _,
-                received: _,
-                required: _,
-                numfailures: _,
-                data_present: _,
-            } => 0x1300,
-            DbError::FunctionFailure {
-                keyspace: _,
-                function: _,
-                arg_types: _,
-            } => 0x1400,
-            DbError::WriteFailure {
-                consistency: _,
-                received: _,
-                required: _,
-                numfailures: _,
-                write_type: _,
-            } => 0x1500,
-            DbError::SyntaxError => 0x2000,
-            DbError::Unauthorized => 0x2100,
-            DbError::Invalid => 0x2200,
-            DbError::ConfigError => 0x2300,
-            DbError::AlreadyExists {
-                keyspace: _,
-                table: _,
-            } => 0x2400,
-            DbError::Unprepared { statement_id: _ } => 0x2500,
-            DbError::Other(code) => *code,
-            DbError::RateLimitReached {
-                op_type: _,
-                rejected_by_coordinator: _,
-            } => protocol_features.rate_limit_error.unwrap(),
-        }
-    }
-
-    /// Decides whether the error can be ignored. If true, the driver can perform
-    /// a speculative retry to the next target.
-    pub fn can_speculative_retry(&self) -> bool {
-        // Do not remove this lint!
-        // It's there for a reason - we don't want new variants
-        // automatically fall under `_` pattern when they are introduced.
-        #[deny(clippy::wildcard_enum_match_arm)]
-        match self {
-            // Errors that will almost certainly appear on other nodes as well
-            DbError::SyntaxError
-            | DbError::Invalid
-            | DbError::AlreadyExists { .. }
-            | DbError::Unauthorized
-            | DbError::ProtocolError => false,
-
-            // Errors that should not appear there - thus, should not be ignored.
-            DbError::AuthenticationError | DbError::Other(_) => false,
-
-            // For now, let's assume that UDF failure is not transient - don't ignore it
-            // TODO: investigate
-            DbError::FunctionFailure { .. } => false,
-
-            // Not sure when these can appear - don't ignore them
-            // TODO: Investigate these errors
-            DbError::ConfigError | DbError::TruncateError => false,
-
-            // Errors that we can ignore and perform a retry on some other node
-            DbError::Unavailable { .. }
-            | DbError::Overloaded
-            | DbError::IsBootstrapping
-            | DbError::ReadTimeout { .. }
-            | DbError::WriteTimeout { .. }
-            | DbError::ReadFailure { .. }
-            | DbError::WriteFailure { .. }
-            // Preparation may succeed on some other node.
-            | DbError::Unprepared { .. }
-            | DbError::ServerError
-            | DbError::RateLimitReached { .. } => true,
-        }
-    }
-}
-
-// Re-exports from scylla-cql-core for backward compatibility.
-pub use scylla_cql_core::frame::response::error::{OperationType, WriteType};
 
 #[cfg(test)]
 mod tests {
